User Preferences GUI implemented and other GUI updates
The framework for the user preferences GUI has been implemented, meaning that the nodes are present and visible but currently have no functionality. This includes the rest of the information needed from the user after application information is entered before the program is able to grab product information and formally build the computer configuration(s).

The GUI itself has also been updated across the board, with margins added wherever necessary.