ComputerBuild Class fleshed out, other changes
There have been changes across the board. The user preferences GUI has been altered and no longer asks for any monitor-related input. This is because reliable access to the data necessary to make this input worthwhile was unable to be found.

The ComputerBuild class has been created and fleshed out.